* Week 3 assignment started;   * BruteCollinearPoints fix double division issue.

+++ src/main/java/org/coursera/princeton/algorithms/week3/Point.java
@@ -82,7 +82,7 @@ public class Point implements Comparable<Point> {
 		} else if (this.y == that.y) {
 			return 0.0d;
 		}
-		return (that.y - this.y) / (that.x - this.x);
+		return (double) (that.y - this.y) / (that.x - this.x);
 	}
 
 	/**
